+++ src/plonemeeting/portal/core/faceted/criteria.py
@@ -15,29 +15,12 @@ class MeetingsCriteria(Criteria):
     def __init__(self, context):
         self.criteria = self.compute_criteria(context)
 
-    def compute_criteria(self, context, must_choose_meeting=True):
-        """ Gets global faceted criteria and unset hidealloption if needed
+    def compute_criteria(self, context):
+        """ Use faceted criteria defined globally
         """
         portal = api.portal.get()
         config_folder = getattr(portal, CONFIG_FOLDER_ID)
         faceted = getattr(config_folder, FACETED_FOLDER_ID)
         self.context = faceted
         criteria = copy.deepcopy(self._criteria())
-        if must_choose_meeting:
-            filtered_criteria = []
-            for criterion in criteria:
-                if criterion.getId() != "annee":
-                    filtered_criteria.append(criterion)
-            return filtered_criteria
-        for criterion in criteria:
-            if criterion.getId() == "seance":
-                criterion.hidealloption = False
         return criteria
-
-
-class ItemsCriteria(MeetingsCriteria):
-    """
-    """
-
-    def __init__(self, context):
-        self.criteria = self.compute_criteria(context, must_choose_meeting=False)