+++ src/plonemeeting/portal/core/setuphandlers.py
@@ -1,30 +1,32 @@
 # -*- coding: utf-8 -*-
-import mimetypes
-
-from Products.CMFPlone.interfaces import INonInstallable
 from imio.helpers.content import richtextval
 from plone import api
 from plone.api import content
+from plone.app.contenttypes.interfaces import IPloneAppContenttypesLayer
+from plone.browserlayer.layer import mark_layer
 from plone.i18n.normalizer.interfaces import IIDNormalizer
 from plone.namedfile.file import NamedFile
+from plonemeeting.portal.core import _
+from plonemeeting.portal.core import logger
+from plonemeeting.portal.core.config import APP_FOLDER_ID
+from plonemeeting.portal.core.config import CONFIG_FOLDER_ID
+from plonemeeting.portal.core.config import FACETED_FOLDER_ID
+from plonemeeting.portal.core.config import FACETED_XML_PATH
+from plonemeeting.portal.core.utils import cleanup_contents
+from plonemeeting.portal.core.utils import create_faceted_folder
+from plonemeeting.portal.core.utils import format_institution_managers_group_id
+from plonemeeting.portal.core.utils import remove_left_portlets
+from plonemeeting.portal.core.utils import remove_right_portlets
+from Products.CMFPlone.interfaces import INonInstallable
 from zope.component import getUtility
 from zope.i18n import translate
 from zope.interface import implementer
+from zope.traversing.interfaces import BeforeTraverseEvent
 
 import dateutil.parser
 import json
+import mimetypes
 import os
-
-from plonemeeting.portal.core import _
-from plonemeeting.portal.core.config import CONFIG_FOLDER_ID
-from plonemeeting.portal.core.config import FACETED_FOLDER_ID
-from plonemeeting.portal.core.utils import (
-    cleanup_contents,
-    format_institution_managers_group_id,
-)
-from plonemeeting.portal.core.utils import create_faceted_folder
-from plonemeeting.portal.core.utils import remove_left_portlets
-from plonemeeting.portal.core.utils import remove_right_portlets
 
 
 @implementer(INonInstallable)
@@ -38,7 +40,6 @@ def post_install(context):
     """Post install script"""
     portal = api.portal.get()
     current_lang = api.portal.get_default_language()[:2]
-    faceted_config = "/faceted/config/items.xml"
 
     if "config" in portal.objectIds():
         return
@@ -64,7 +65,8 @@ def post_install(context):
     )
     subtyper = faceted.restrictedTraverse("@@faceted_subtyper")
     subtyper.enable()
-    with open(os.path.dirname(__file__) + faceted_config, "rb") as faceted_config:
+    faceted_config_path = os.path.join(os.path.dirname(__file__), FACETED_XML_PATH)
+    with open(faceted_config_path, "rb") as faceted_config:
         faceted.unrestrictedTraverse("@@faceted_exportimport").import_xml(
             import_file=faceted_config
         )
@@ -94,6 +96,18 @@ def create_demo_content(context):
     :param context:
     """
     portal = api.portal.get()
+    request = portal.REQUEST
+    # make sure the plone.app.contenttypes BrowserLayer is enabled
+    # because it is needed to create demo content (to get @@file_view)
+    # when demo content added at Plone Site creation time
+    # the plone.app.contenttypes BrowserLayer is not enabled
+    if not IPloneAppContenttypesLayer.providedBy(request):
+        logger.warn("IPloneAppContenttypesLayer not enable on REQUEST, enabling it.")
+        event = BeforeTraverseEvent(portal, request)
+        mark_layer(portal, event)
+    else:
+        logger.info("IPloneAppContenttypesLayer already enabled on REQUEST.")
+
     current_dir = os.path.abspath(os.path.dirname(__file__))
     json_path = os.path.join(current_dir, "profiles/demo/data/data.json")
 
@@ -111,6 +125,7 @@ def create_demo_content(context):
                 type="Institution",
                 id=institution_id,
                 title=institution["title"],
+                categories_mappings=institution["categories_mappings"],
                 representatives_mappings=institution["representatives_mappings"],
                 plonemeeting_url=institution["plonemeeting_url"],
                 username=institution["username"],
@@ -138,7 +153,7 @@ def create_demo_content(context):
             group = api.group.get(format_institution_managers_group_id(institution_obj))
             group.addMember(user.id)
 
-            for meeting in institution["meetings"]:
+            for meeting in institution[APP_FOLDER_ID]:
                 date_time = dateutil.parser.parse(meeting["datetime"])
                 meeting_obj = content.create(
                     container=institution_obj,
@@ -167,7 +182,7 @@ def create_demo_content(context):
                             meeting["plonemeeting_last_modified"]
                         ),
                     )
-
+                    item_obj.formatted_title = richtextval("<p>" + item_obj.title + "</p>")
                     if "files" in item:
                         for file in item["files"]:
                             create_file(item_obj, file)