+++ src/plonemeeting/portal/core/tests/test_vocabularies.py
@@ -32,9 +32,34 @@ class TestVocabularies(PmPortalDemoFunctionalTestCase):
         vocab = queryUtility(
             IVocabularyFactory, "plonemeeting.portal.vocabularies.representatives"
         )
-        values_city1 = vocab(self.item)
-        values_city2 = vocab(self.item2)
-        self.assertFalse(len(values_city1) == len(values_city2))
+        voc = vocab(self.item)
+        values_city1 = [term.title for term in voc._terms]
+        self.assertListEqual(values_city1,
+                             ['Mr DUPONT', 'Mr Dupuis', 'Mr Oniz', 'Mr Baka', 'Mr Kuro'])
+        voc = vocab(self.item2)
+        values_city2 = [term.title for term in voc._terms]
+        self.assertListEqual(values_city2,
+                             ['Mme LOREM', 'Mme Ipsum', 'Mr Wara', 'Mr Bara'])
+
+    def testLongRepresentativesVocabulary(self):
+        vocab = queryUtility(
+            IVocabularyFactory, "plonemeeting.portal.vocabularies.long_representatives"
+        )
+        voc = vocab(self.item)
+        values_city1 = [term.title for term in voc._terms]
+        self.assertListEqual(values_city1,
+                             ['Mr DUPONT Bourgmestre F.F.',
+                              'Mr Dupuis 1ère Échevin',
+                              "Mr Oniz, Échevin de l'éducation",
+                              "Mr Baka, Échevin de des sports",
+                              "Mr Kuro, Échevin de la culture"])
+        voc = vocab(self.item2)
+        values_city2 = [term.title for term in voc._terms]
+        self.assertListEqual(values_city2,
+                             ['Mme LOREM Bourgmestre',
+                              'Mme Ipsum 1ère Échevine',
+                              'Mr Wara, Échevin du tourisme',
+                              'Mr Bara, Échevin du Développement économique'])
 
     def testMeetingDatesVocabulary(self):
         vocab = queryUtility(